Move deepupdate to utils
